:construction: tweak system_prompt
